+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @drupal-canvas/cli
 
+## 0.21.2
+
+### Patch Changes
+
+- cbb1a53: Fix content template creation by omitting the unsupported `label`
+  property from create requests.
+
 ## 0.21.1
 
 ### Patch Changes